LayoutTests/platform/chromium-mac/fast/js/delete-getters-setters-expected.txt

This test checks that deletion of properties works properly with getters and setters.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS b1.property threw exception TypeError: Cannot read property 'property' of undefined.
PASS a2.property threw exception TypeError: Cannot read property 'property' of undefined.
PASS b3.property threw exception TypeError: Cannot read property 'property' of undefined.
PASS a4.property threw exception TypeError: Cannot read property 'property' of undefined.
PASS b5.property threw exception TypeError: Cannot read property 'property' of undefined.
PASS a6.property threw exception TypeError: Cannot read property 'property' of undefined.
PASS b7.property threw exception TypeError: Cannot read property 'property' of undefined.
PASS a8.property threw exception TypeError: Cannot read property 'property' of undefined.
PASS o1.b.property threw exception TypeError: Cannot read property 'property' of undefined.
PASS o1.a.property threw exception TypeError: Cannot read property 'property' of undefined.
PASS o3.b.property threw exception TypeError: Cannot read property 'property' of undefined.
PASS o4.a.property threw exception TypeError: Cannot read property 'property' of undefined.
PASS o5.b.property threw exception TypeError: Cannot read property 'property' of undefined.
PASS o6.a.property threw exception TypeError: Cannot read property 'property' of undefined.
PASS o7.b.property threw exception TypeError: Cannot read property 'property' of undefined.
PASS o8.a.property threw exception TypeError: Cannot read property 'property' of undefined.
PASS successfullyParsed is true

TEST COMPLETE
